Deployment/requirements.txt
streamlit
pandas
numpy
scipy
tensorflow-cpu
matplotlib
pickle-mixin
scikit-learn==1.0.2